change in my praposal page and add hours and day ago functionality on view proposal page
